Tweak and refactor neural nets and losses.
Detailed changes are as follows:

- add custom final activation to MLPClassifier
- update license information related to using sparsemax code from TFA
- refactor sparsemax functions
- refactor some loss functions and the folder structure
- update dependencies
- update README
